verifier on jolt

+++ tracer/src/instruction/and.rs
@@ -17,8 +17,8 @@ declare_riscv_instr!(
 
 impl AND {
     fn exec(&self, cpu: &mut Cpu, _: &mut <AND as RISCVInstruction>::RAMAccess) {
-        cpu.x[self.operands.rd] =
-            cpu.sign_extend(cpu.x[self.operands.rs1] & cpu.x[self.operands.rs2]);
+        cpu.x[self.operands.rd as usize] =
+            cpu.sign_extend(cpu.x[self.operands.rs1 as usize] & cpu.x[self.operands.rs2 as usize]);
     }
 }
