+++ tracer/src/instruction/auipc.rs
@@ -17,7 +17,7 @@ declare_riscv_instr!(
 
 impl AUIPC {
     fn exec(&self, cpu: &mut Cpu, _: &mut <AUIPC as RISCVInstruction>::RAMAccess) {
-        cpu.x[self.operands.rd] =
+        cpu.x[self.operands.rd as usize] =
             cpu.sign_extend(self.address as i64 + normalize_imm(self.operands.imm));
     }
 }
