+++ tracer/src/instruction/srli.rs
@@ -27,8 +27,8 @@ impl SRLI {
             Xlen::Bit32 => 0x1f,
             Xlen::Bit64 => 0x3f,
         };
-        cpu.x[self.operands.rd] = cpu.sign_extend(
-            cpu.unsigned_data(cpu.x[self.operands.rs1])
+        cpu.x[self.operands.rd as usize] = cpu.sign_extend(
+            cpu.unsigned_data(cpu.x[self.operands.rs1 as usize])
                 .wrapping_shr(self.operands.imm as u32 & mask) as i64,
         );
     }
